change team names 2
This commit reflects the team names changes on the UI. Now you can switch teams through the main menu.
I changed the team commands to more neutral "red" and "blue" so they are clear to anyone and can stay the same even if we decided to change team names again sometime in the future.
The new team names are "Aesir" and "Vanir", two factions in norse mythology. not that it makes much sense to have these two teams fighting in Valhalla, but still it's a nice reference to norse mythology.