adjust chat commands
while triggering chat with the RETURN key is modern, the fact remains that this is more of the case on games that either don't have multiple chat channels, or a chat widget that allows easily changing between their multitude of chat channels. plenty of games still use T and Y for chat, so let's stick with what people are familiar with.

also fixes the team chat not working in non-team modes, as it now defaults to global chat