implement texintmul
it's an int-based version of texmad, but without the add part of the equation -- easier to work with in a hex-based color system, whereas with mad we need to convert from int to rgb on the cubescript end and then to 0-1f range before we can even pass it on. this approach will be inherently faster as a result, even if it's a single-run situation. we'll also have less of a need to cache the results ahead of time for efficiency on the cubescript side, meaning cleaner overall code.