implement ID system for sleep calls
sleep calls with an ID given to them are uninterruptible. They can only be removed if explicitly calling `delsleep` with a matching ID, or by expiring naturally once the expected time elapses. Additionally, `findsleep` was added to return true/false if a particular sleep call matching the given ID is found active.

Lastly, added a new execident on clearsleep so that its use can be detected.