New trigger system
Entities have a new field: `label`.
The label is interpreted as a list of space-separated words.
Entities can be queried by their label: a label matches a query string if every word of the query string is present in the label. This allows labels to work similarly to classes in HTML/CSS.

Triggers are objects that emit events when certain conditions apply. Events can be caught by handler functions to run custom cubescript. A single event can be caught by multiple event handlers.

Trigger types and the events they emit:
- Item: an item that can be picked up. Emits "Proximity" once when the player gets inside the ent's radius, and disables itself immediately.
- UsableItem: an item that can be used by pressing the "Use item" key. Emits "Proximity" whenever the player is close and does not disable itself. When the player presses the "Use item" key, emits "Use" and disables itself.
- Marker: disabled by default; when enabled, marks its location on the hud. Emits the same events as "Item".

Additionally, all triggers that were active when the player entered their radius emit a "Distance" event once when the player dies or moves outside of the ent's exit radius. This occurs regardless of the current activation state.

In addition to the ones mentioned above, there is also a "Manual" event type, which can only be emitted from cubescript.

Handling events:
Event handlers are defined like this:
```
trigger "query string" "event type" [ code to run ]
```
When a trigger whose label matches the query string emits an event of the specified type, the code is executed. A single event may fire more than one event handler. Inside the event handler, the `curtrigger` variable can be used to access the ent id of the trigger that emitted the event. The event type must be lowercase.

Other stuff:
- `matchentquery "query" "label"` can be used to test if a label matches a query
- `entquery "type" "query"` returns a list of the id's of all ents of the specified type whose label matches the query
- `emittriggerevent id "event type"` emits an event from the specified ent id
- `cleartriggereventhandlers` removes all the currently specified event handlers
- `useitem`, bound to the "B" key by default, attempts to use a nearby item. It cannot be used from the command line.
- `setrespawnpoint id` sets the respawn point to the specified ent id. The ent's attr1 will be the player's yaw at respawn.
- `triggerstate id` and `triggerstate id state` can be used respectively to get and set a trigger's activation state
- `dotriggerpickupevents cn` plays the "pickup" screen effect. It was previously performed automatically when a trigger fired. The cn is currently useless and should be omitted.
- `playtriggersound ent_id sound_id cn` plays a sound at the location of the ent
- `triggertoggle` and `triggerswap` have been removed
- `mapmodeltriggerstate id` returns the current trigger state of a mapmodel

Breaks the tutorial
Bumps the map format version
Bumps the net protocol version