Write ent activity state to disk
A list of disabled entities is written to map files, allowing entities to be disabled at map load.

This data is stored in the `extras` field of map files.
The `extras` field, previously unused, now stores data as an array of "blocks", composed of:
- int tag
- int size
- uchar[] data
Where `tag` specifies the type of data contained in the block.